pyproject.toml
[tool.poetry]
name = "domainspotter"
version = "0.1.0"
description = ""
authors = ["erik aronesty <erik@q32.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
openapi = "^2.0.0"
aiodns = "^3.1.1"
pydantic = "^2.6.1"
asyncpg = "^0.30.0"
httpx = "^0.28.1"
asgi-lifespan = "^2.1.0"
python-dotenv = "^1.1.0"
openai = "^1.78.1"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
jinja2 = "^3.1.6"
async-lru = "^2.0.5"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.1.1"
pytest-xdist = "^3.6.1"
pytest-asyncio = "^0.23.5"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
